+++ src/main/kotlin/com/jakewharton/testdistribution/TestDistributionPlugin.kt
@@ -1,8 +1,13 @@
 package com.jakewharton.testdistribution
 
+import com.android.build.api.artifact.ScopedArtifact
+import com.android.build.api.variant.AndroidComponentsExtension
+import com.android.build.api.variant.HasHostTests
+import com.android.build.api.variant.ScopedArtifacts
 import java.io.File
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.file.RegularFile
 import org.gradle.api.plugins.BasePluginExtension
 import org.gradle.api.tasks.Copy
 import org.gradle.api.tasks.application.CreateStartScripts
@@ -34,6 +39,93 @@ public class TestDistributionPlugin : Plugin<Project> {
 					"project ${project.path}"
 				}
 				"No compatible language plugin applied for JVM test distribution ($name)"
+			}
+		}
+
+		project.pluginManager.withPlugin("com.android.library") {
+			gotPlugin = true
+
+			val base = project.extensions.getByType(BasePluginExtension::class.java)
+
+			val androidComponents = project.extensions.getByType(AndroidComponentsExtension::class.java)
+			androidComponents.onVariants { variant ->
+				if (variant !is HasHostTests) return@onVariants
+
+				val variantUpperName = variant.name.replaceFirstChar(Char::uppercase)
+				for ((testUpperName, hostTest) in variant.hostTests) {
+					val testName = testUpperName.replaceFirstChar(Char::lowercase)
+					val name = variant.name + testUpperName
+					val nameUpper = variantUpperName + testUpperName
+
+					val dummyClassesProvider = project.tasks.register("internal${nameUpper}Classes", ScopedArtifactDummyTask::class.java)
+					hostTest.artifacts
+						.forScope(ScopedArtifacts.Scope.ALL)
+						.use(dummyClassesProvider)
+						.toGet(
+							ScopedArtifact.CLASSES,
+							ScopedArtifactDummyTask::jars,
+							ScopedArtifactDummyTask::classDirs,
+						)
+
+					val testJarProvider = project.tasks.register("jar$nameUpper", Jar::class.java) {
+						it.from(dummyClassesProvider.flatMap(ScopedArtifactDummyTask::classDirs))
+						it.archiveAppendix.set(variant.name)
+						it.archiveClassifier.set(testName)
+					}
+
+					val testScriptsProvider = project.tasks.register("scripts$nameUpper", CreateStartScripts::class.java) { task ->
+						task.outputDir = project.layout.buildDirectory.dir("scripts/$nameUpper").get().asFile
+						task.applicationName = base.archivesName.get() + "-test"
+
+						val classpath = project.objects.fileCollection()
+						classpath.from(testJarProvider.map { it.outputs.files })
+						classpath.from(dummyClassesProvider.flatMap { it.jars })
+
+						task.classpath = classpath
+
+						task.mainClass.set(
+							dummyClassesProvider.flatMap {
+								it.classDirs.zip(it.jars) { classDirs, jars ->
+									val testClasses = project.objects.fileTree()
+									for (classDir in classDirs) {
+										testClasses.from(classDir)
+									}
+
+									val testFqcns = gradleSupport.detectTestClassNames(
+										testClasses,
+										testClasses.files.toList(),
+										jars.map(RegularFile::getAsFile),
+									)
+									"org.junit.runner.JUnitCore ${testFqcns.joinToString(" ") { """"$it"""" }}"
+								}
+							},
+						)
+					}
+
+					val installProvider = project.tasks.register("install${nameUpper}Distribution", Copy::class.java) {
+						it.group = "distribution"
+						it.description = "Installs $name as a distribution as-is."
+
+						it.into("bin") {
+							it.from(testScriptsProvider)
+						}
+						it.into("lib") {
+							it.from(testJarProvider)
+							it.from(dummyClassesProvider.flatMap { it.jars })
+						}
+						it.destinationDir = project.layout.buildDirectory.dir("install/$name").get().asFile
+					}
+
+					project.tasks.register("zip${nameUpper}Distribution", Zip::class.java) {
+						it.group = "distribution"
+						it.description = "Bundles $name as a distribution."
+
+						it.from(installProvider)
+						it.destinationDirectory.set(project.layout.buildDirectory.dir("dist"))
+						it.archiveAppendix.set(variant.name)
+						it.archiveClassifier.set(testName)
+					}
+				}
 			}
 		}